修复卸载全息图报错

+++ src/main/java/com/mythmc/externs/hook/hologram/DecentHologramsHook.java
@@ -70,9 +70,12 @@ public class DecentHologramsHook {
     }
 
     public static void removeHologram() {
-        // 清空缓存
-        Hologram hologram = DHAPI.getHologram("minebbstopper_rank");
-        if (hologram != null)
-            hologram.delete();
+        if (setupDecentHolograms() && HologramFile.hologramEnable) {
+            // 清空缓存
+            Hologram hologram = DHAPI.getHologram("minebbstopper_rank");
+            if (hologram != null)
+                hologram.delete();
+        }
     }
+
 }
